+++ pymince/dictionary.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+
 """Useful functions that use dictionaries."""
 
 import collections