+++ tests/test_json/test_dump_into.py
@@ -7,6 +7,8 @@ import pytest
 
 import pymince.json
 
+EXTENSIONS = (".json.gz", ".json.bz2", ".json.xz", ".json")
+
 
 def test_dumped_file():
     with tempfile.TemporaryDirectory() as tmpdir:
@@ -15,7 +17,7 @@ def test_dumped_file():
         assert os.path.exists(filename)
 
 
-@pytest.mark.parametrize("extension", pymince.json.EXTENSIONS)
+@pytest.mark.parametrize("extension", EXTENSIONS)
 def test_dumped_data(extension):
     data = {"key": "ñó", "nested": [1, 2, 3]}
     with tempfile.TemporaryDirectory() as tmpdir: