+++ tests/test_json/test_load_from.py
@@ -7,14 +7,18 @@ import tempfile
 import pytest
 
 import pymince.json
+import pymince._constants
+import pymince.file
+
+EXTENSIONS = (".json.gz", ".json.bz2", ".json.xz", ".json")
 
 
-@pytest.mark.parametrize("extension", pymince.json.EXTENSIONS)
+@pytest.mark.parametrize("extension", EXTENSIONS)
 def test_load_given_filepath(extension):
     expected = data = {"key": "ñó", "nested": [1, 2, 3]}
     with tempfile.TemporaryDirectory() as tmp:
         path = os.path.join(tmp, f"foo{extension}")
-        with pymince.json.xopen(path, "wt", pymince.json.ENCODING) as fd:
+        with pymince.file.xopen(path, mode="wt", encoding=pymince._constants.utf_8) as fd:
             fd.write(json.dumps(data))
         result = pymince.json.load_from(path)
     assert result == expected